+++ mongock-spring/src/test/java/com/github/cloudyrock/spring/test/changelogs/withChangockAnnotations/ChangeLogwithChangockAnnotations.java
@@ -1,0 +1,19 @@
+package com.github.cloudyrock.spring.test.changelogs.withChangockAnnotations;
+
+import io.changock.migration.api.annotations.ChangeLog;
+import io.changock.migration.api.annotations.ChangeSet;
+
+/**
+ *
+ * @since 27/07/2014
+ */
+@ChangeLog(order = "1")
+public class ChangeLogwithChangockAnnotations {
+
+  @ChangeSet(author = "testuser", id = "withChangockAnnotations", order = "01", systemVersion = "0.1")
+  public void withChangockAnnotations() {
+    System.out.println("invoked 1");
+  }
+
+
+}